Robot: change to using the Adafruit PCA9685 PWM driver and adjust pan/tilt
So in amongst all the attempts to get a nicer power supply for the servos, I managed
to burn out the Emakefun I2C board I was using. I'd been thinking of switching to the
Adafruit one for a while, but the Emakefun was working and all that...until it wasn't.
Nice thing about the Adafruit is the external power rail (with a nice big capacitor),
which lets me power the servos from a 5V BEC instead of drawing from the Hercules
controller board (which is only supposed to provide ~1A)...